fix(palette): keep github_bridge.rs under the monolith line cap
Task 4's Feed wiring pushed github_bridge.rs to 504 lines (over the
500-line cap). Moves build_feed_payload/github_browse_feed into
github_feed.rs (Feed-specific dispatch, not shared with the other four
GitHubRequestKinds) per the plan's Task 9 Step 7 guidance, widening the
now-cross-module GitHubBrowseRequest/GitHubBrowseResult fields and a
few bridge helpers to pub(crate). Also fixes two clippy -D warnings
findings (unnecessary_sort_by, dead FeedBadge::Diff variant) and three
biome lint warnings (array-index key, non-null assertion, missing
useEffect dep) surfaced during the full verification pass.
